ParticleUtil 및 SoundUtil의 전투 시스템에서 사용하는 메소드를 CombatUtil로 이동
