+++ src/main/java/com/dace/dmgr/combat/character/jager/action/JagerA2Entity.java
@@ -175,7 +175,7 @@ public final class JagerA2Entity extends SummonEntity<MagmaCube> implements HasR
     @Override
     public void onDamage(Attacker attacker, int damage, int reducedDamage, @NonNull DamageType damageType, Location location, boolean isCrit, boolean isUlt) {
         SoundUtil.playNamedSound(NamedSound.COMBAT_JAGER_A2_DAMAGE, entity.getLocation(), 1 + damage * 0.001);
-        ParticleUtil.playBreakEffect(location, entity, damage);
+        CombatUtil.playBreakEffect(location, entity, damage);
     }
 
     @Override
