+++ src/main/java/com/dace/dmgr/combat/character/quaker/Quaker.java
@@ -1,5 +1,6 @@
 package com.dace.dmgr.combat.character.quaker;
 
+import com.dace.dmgr.combat.CombatUtil;
 import com.dace.dmgr.combat.DamageType;
 import com.dace.dmgr.combat.action.info.ActiveSkillInfo;
 import com.dace.dmgr.combat.action.info.PassiveSkillInfo;
@@ -9,7 +10,6 @@ import com.dace.dmgr.combat.character.quaker.action.*;
 import com.dace.dmgr.combat.entity.Attacker;
 import com.dace.dmgr.combat.entity.CombatUser;
 import com.dace.dmgr.util.NamedSound;
-import com.dace.dmgr.util.ParticleUtil;
 import com.dace.dmgr.util.SoundUtil;
 import com.dace.dmgr.util.StringFormUtil;
 import lombok.Getter;
@@ -68,7 +68,7 @@ public final class Quaker extends Character {
 
     @Override
     public void onDamage(@NonNull CombatUser victim, Attacker attacker, int damage, @NonNull DamageType damageType, Location location, boolean isCrit) {
-        ParticleUtil.playBleedingEffect(location, victim.getEntity(), damage);
+        CombatUtil.playBleedingEffect(location, victim.getEntity(), damage);
     }
 
     @Override
