+++ src/main/java/com/dace/dmgr/util/ParticleUtil.java
@@ -8,10 +8,7 @@ import lombok.experimental.UtilityClass;
 import org.bukkit.Location;
 import org.bukkit.Material;
 import org.bukkit.Particle;
-import org.bukkit.block.Block;
-import org.bukkit.entity.Entity;
 import org.bukkit.entity.Player;
-import org.jetbrains.annotations.Nullable;
 
 /**
  * 입자 생성 기능을 제공하는 클래스.
@@ -166,60 +163,6 @@ public final class ParticleUtil {
                 offsetY, offsetZ, (float) speed);
 
         packet.sendPacket(player);
-    }
-
-    /**
-     * 지정한 위치 또는 엔티티에 출혈 효과를 재생한다.
-     *
-     * @param location 대상 위치
-     * @param entity   대상 엔티티
-     * @param damage   피해량
-     */
-    public static void playBleedingEffect(@Nullable Location location, @Nullable Entity entity, int damage) {
-        if (location == null && entity == null)
-            return;
-
-        if (location == null)
-            ParticleUtil.playBlock(ParticleUtil.BlockParticle.BLOCK_DUST, Material.REDSTONE_BLOCK, 0,
-                    entity.getLocation().add(0, entity.getHeight() / 2, 0), damage == 0 ? 1 : (int) Math.ceil(damage * 0.1),
-                    entity.getWidth() / 4, entity.getHeight() / 4, entity.getWidth() / 4, damage == 0 ? 0.03 : 0.1);
-        else
-            ParticleUtil.playBlock(ParticleUtil.BlockParticle.BLOCK_DUST, Material.REDSTONE_BLOCK, 0, location, (int) Math.ceil(damage * 0.06),
-                    0, 0, 0, 0.1);
-    }
-
-    /**
-     * 지정한 위치 또는 엔티티에 파괴 효과를 재생한다.
-     *
-     * @param location 대상 위치
-     * @param entity   대상 엔티티
-     * @param damage   피해량
-     */
-    public static void playBreakEffect(@Nullable Location location, @Nullable Entity entity, int damage) {
-        if (location == null && entity == null)
-            return;
-
-        if (location == null)
-            ParticleUtil.playBlock(ParticleUtil.BlockParticle.BLOCK_DUST, Material.IRON_BLOCK, 0,
-                    entity.getLocation().add(0, entity.getHeight() / 2, 0), damage == 0 ? 1 : (int) Math.ceil(damage * 0.07),
-                    entity.getWidth() / 4, entity.getHeight() / 4, entity.getWidth() / 4, damage == 0 ? 0.03 : 0.1);
-        else
-            ParticleUtil.playBlock(ParticleUtil.BlockParticle.BLOCK_DUST, Material.IRON_BLOCK, 0, location, (int) Math.ceil(damage * 0.04),
-                    0, 0, 0, 0.1);
-    }
-
-    /**
-     * 지정한 위치에 블록 타격 효과를 재생한다.
-     *
-     * @param location        대상 위치
-     * @param block           블록
-     * @param scaleMultiplier 규모(입자의 양 및 범위) 배수
-     */
-    public static void playBlockHitEffect(@NonNull Location location, @NonNull Block block, double scaleMultiplier) {
-        ParticleUtil.playBlock(ParticleUtil.BlockParticle.BLOCK_DUST, block.getType(), block.getData(), location,
-                (int) (6 * scaleMultiplier), 0.06 * scaleMultiplier, 0.06 * scaleMultiplier, 0.06 * scaleMultiplier, 0.1);
-        ParticleUtil.play(Particle.TOWN_AURA, location, (int) (25 * scaleMultiplier),
-                0.05 * scaleMultiplier, 0.05 * scaleMultiplier, 0.05 * scaleMultiplier, 0);
     }
 
     /**
